Added memcheck option to run Valgrind checks
Adding a --memcheck option to `./bin/test` will
now cause Valgrind's `memcheck` test to run over
the unit_tests binary.

+++ build.sh
@@ -12,7 +12,9 @@ BUILD=${1:-Debug}
 mkdir -p ${BUILDDIR}
 cd ${BUILDDIR}
 
-conan install ${BASEDIR} -if=${BUILDDIR} -pr=default --build=missing
+if [[ -f ${BASEDIR}/conanfile.txt ]]; then
+  conan install ${BASEDIR} -if=${BUILDDIR} -pr=default --build=missing
+fi
 
 cmake -DCMAKE_CXX_COMPILER=${CLANG} \
       -DCMAKE_BUILD_TYPE=${BUILD} \
